+++ ignite/pkg/cosmosgen/generate_openapi.go
@@ -5,19 +5,16 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 
 	"github.com/blang/semver/v4"
 	"github.com/iancoleman/strcase"
 
 	"github.com/ignite/cli/v28/ignite/pkg/cache"
-	"github.com/ignite/cli/v28/ignite/pkg/cosmosanalysis/module"
 	"github.com/ignite/cli/v28/ignite/pkg/cosmosbuf"
 	"github.com/ignite/cli/v28/ignite/pkg/dirchange"
 	"github.com/ignite/cli/v28/ignite/pkg/errors"
-	"github.com/ignite/cli/v28/ignite/pkg/nodetime"
-	swaggercombine "github.com/ignite/cli/v28/ignite/pkg/nodetime/programs/swagger-combine"
+	swaggercombine "github.com/ignite/cli/v28/ignite/pkg/swagger-combine"
 	"github.com/ignite/cli/v28/ignite/pkg/xos"
 )
 
@@ -30,25 +27,11 @@ func (g *generator) openAPITemplate() string {
 	return filepath.Join(g.appPath, g.protoDir, "buf.gen.swagger.yaml")
 }
 
-func (g *generator) openAPITemplateForSTA() string {
-	return filepath.Join(g.appPath, g.protoDir, "buf.gen.sta.yaml")
-}
-
 func (g *generator) generateOpenAPISpec(ctx context.Context) error {
 	var (
 		specDirs []string
-		conf     = swaggercombine.Config{
-			Swagger: "2.0",
-			Info: swaggercombine.Info{
-				Title: "HTTP API Console",
-			},
-		}
+		conf     = swaggercombine.New("HTTP API Console", g.goModPath)
 	)
-	command, cleanup, err := nodetime.Command(nodetime.CommandSwaggerCombine)
-	if err != nil {
-		return err
-	}
-	defer cleanup()
 	defer func() {
 		for _, dir := range specDirs {
 			os.RemoveAll(dir)
@@ -85,7 +68,7 @@ func (g *generator) generateOpenAPISpec(ctx context.Context) error {
 			if err := os.WriteFile(specPath, existingSpec, 0o644); err != nil {
 				return err
 			}
-			return conf.AddSpec(name, specPath)
+			return conf.AddSpec(name, specPath, true)
 		}
 
 		hasAnySpecChanged = true
@@ -122,7 +105,7 @@ func (g *generator) generateOpenAPISpec(ctx context.Context) error {
 			if err := specCache.Put(cacheKey, f); err != nil {
 				return err
 			}
-			if err := conf.AddSpec(name, spec); err != nil {
+			if err := conf.AddSpec(name, spec, true); err != nil {
 				return err
 			}
 		}
@@ -174,101 +157,12 @@ func (g *generator) generateOpenAPISpec(ctx context.Context) error {
 		}
 	}
 
-	sort.Slice(conf.APIs, func(a, b int) bool { return conf.APIs[a].ID < conf.APIs[b].ID })
-
-	// ensure out dir exists.
-	outDir := filepath.Dir(out)
-	if err := os.MkdirAll(outDir, 0o766); err != nil {
-		return err
-	}
-
 	// combine specs into one and save to out.
-	if err := swaggercombine.Combine(ctx, conf, command, out); err != nil {
+	if err := conf.Combine(out); err != nil {
 		return err
 	}
 
 	return dirchange.SaveDirChecksum(specCache, out, g.appPath, out)
-}
-
-func (g *generator) generateModuleOpenAPISpec(ctx context.Context, m module.Module, out string) error {
-	var (
-		specDirs []string
-		conf     = swaggercombine.Config{
-			Swagger: "2.0",
-			Info: swaggercombine.Info{
-				Title: "HTTP API Console " + m.Pkg.Name,
-			},
-		}
-	)
-	command, cleanup, err := nodetime.Command(nodetime.CommandSwaggerCombine)
-	if err != nil {
-		return err
-	}
-	defer cleanup()
-
-	defer func() {
-		for _, dir := range specDirs {
-			os.RemoveAll(dir)
-		}
-	}()
-
-	// gen generates a spec for a module where it's source code resides at src.
-	// and adds needed swaggercombine configure for it.
-	gen := func(m module.Module) (err error) {
-		dir, err := os.MkdirTemp("", "gen-openapi-module-spec")
-		if err != nil {
-			return err
-		}
-
-		err = g.buf.Generate(ctx, m.Pkg.Path, dir, g.openAPITemplateForSTA(), cosmosbuf.ExcludeFiles("module.proto"))
-		if err != nil {
-			return err
-		}
-
-		specs, err := xos.FindFilesExtension(dir, xos.JSONFile)
-		if err != nil {
-			return err
-		}
-
-		for _, spec := range specs {
-			if err != nil {
-				return err
-			}
-			if err := conf.AddSpec(strcase.ToCamel(m.Pkg.Name), spec); err != nil {
-				return err
-			}
-		}
-		specDirs = append(specDirs, dir)
-
-		return nil
-	}
-
-	// generate specs for each module and persist them in the file system
-	// after add their path and config to swaggercombine.Config so we can combine them
-	// into a single spec.
-	add := func(modules []module.Module) error {
-		for _, m := range modules {
-			if err := gen(m); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-
-	// protoc openapi generator acts weird on concurrent run, so do not use goroutines here.
-	if err := add([]module.Module{m}); err != nil {
-		return err
-	}
-
-	sort.Slice(conf.APIs, func(a, b int) bool { return conf.APIs[a].ID < conf.APIs[b].ID })
-
-	// ensure out dir exists.
-	outDir := filepath.Dir(out)
-	if err := os.MkdirAll(outDir, 0o766); err != nil {
-		return err
-	}
-	// combine specs into one and save to out.
-	return swaggercombine.Combine(ctx, conf, command, out)
 }
 
 func extractRootModulePath(fullPath string) string {