+++ changelog.md
@@ -25,6 +25,7 @@
 - [#4194](https://github.com/ignite/cli/pull/4194) Bump client/v2 to `v2.0.0-beta.2`
   If you are uprading manually, check out the recommended changes in `root.go` from the above PR.
 - [#4210](https://github.com/ignite/cli/pull/4210) Improve default home wiring
+- [#4077](https://github.com/ignite/cli/pull/4077) Merge the swagger files manually instead use nodetime `swagger-combine`
 
 ### Fixes